Partially taken from the next branch. Created/refactored interfaces and custom exceptions

+++ src/main/java/telran/game/bulls_cows/GamerGame.java
@@ -3,4 +3,4 @@ package telran.game.bulls_cows;
 public record GamerGame(
         int gamer4game_id,
         String gamer_name,
-        int game_id) { }
+        Long game_id) { }
